backend/requirements.txt
asgiref==3.5.2
distlib==0.3.6
dj-database-url==1.0.0
Django==4.1.2
django-cors-headers==3.13.0
django-heroku==0.3.1
djangorestframework==3.14.0
filelock==3.8.0
gunicorn==20.1.0
joblib==1.2.0
numpy==1.23.3
pandas==1.5.0
Pillow==9.2.0
platformdirs==2.5.2
psycopg2==2.9.4
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
scikit-learn==1.0.2
scipy==1.9.2
six==1.16.0
sqlparse==0.4.3
threadpoolctl==3.1.0
tzdata==2022.4
virtualenv==20.16.5
whitenoise==6.2.0
